Move TTS toggle into sound options menu
