+++ src/client/java/com/owlmaddie/mixin/client/GameOptionsScreenAccessor.java
@@ -1,0 +1,12 @@
+package com.owlmaddie.mixin.client;
+
+import net.minecraft.client.gui.screen.option.GameOptionsScreen;
+import net.minecraft.client.gui.widget.OptionListWidget;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(GameOptionsScreen.class)
+public interface GameOptionsScreenAccessor {
+    @Accessor("body")
+    OptionListWidget creaturechat$getBody();
+}
